FE - Refactor general skeleton moved types.ts and globalFunctions to utils.
